src/main/java/bootcamp/transactionmicroservice/infrastructure/exceptionhandler/GlobalExceptionHandler.java
package bootcamp.transactionmicroservice.infrastructure.exceptionhandler;

import bootcamp.transactionmicroservice.domain.exceptions.SupplyNotFoundException;
import bootcamp.transactionmicroservice.domain.exceptions.SupplyValueException;
import org.springframework.http.HttpHeaders;
import org.springframework.http.HttpStatus;
import org.springframework.http.HttpStatusCode;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.MethodArgumentNotValidException;
import org.springframework.web.bind.annotation.ExceptionHandler;
import org.springframework.web.bind.annotation.RestControllerAdvice;
import org.springframework.web.context.request.WebRequest;
import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;

import java.util.HashMap;
import java.util.Map;

@RestControllerAdvice
public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {

    @ExceptionHandler(SupplyNotFoundException.class)
    public String handleSupplyNotFoundException(SupplyNotFoundException e) {
        return e.getMessage();
    }

    @ExceptionHandler(SupplyValueException.class)
    public String handleSupplyValueException(SupplyValueException e) {
        return e.getMessage();
    }
    @Override
    protected ResponseEntity<Object> handleMethodArgumentNotValid(MethodArgumentNotValidException ex, HttpHeaders headers, HttpStatusCode status, WebRequest request) {
        Map<String,Object> errors = new HashMap<>();
        ex.getBindingResult().getFieldErrors().forEach(error ->

                errors.put(error.getField(),error.getDefaultMessage())
        );

        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errors);
    }
}
